Split initializing routine from the host model
